Fixed exceptions? and added Handlers

+++ server/src/main/java/service/UserService.java
@@ -1,8 +1,8 @@
 package service;
 
 import dataaccess.AuthDao;
-import dataaccess.DataAccessException;
 import dataaccess.UserDao;
+import dataaccess.exceptions.ResponseException;
 import model.AuthData;
 import model.UserData;
 
@@ -15,44 +15,50 @@ public class UserService {
         this.authDao = authDao;
     }
 
-    public AuthData register(UserData user) throws DataAccessException {
+    public AuthData register(UserData user) throws ResponseException {
         if (user == null || user.username() == null || user.username().isEmpty() || user.password() == null || user.password().isEmpty() || user.email() == null || user.email().isEmpty()){
-            throw new DataAccessException("Invalid data provided");
+            throw new ResponseException(400, "Error: Bad Request");
+        }
+
+        UserData currentUser = userDao.getUser(user.username());
+        if (currentUser != null){
+            throw new ResponseException(403, "Error: already taken");
         }
 
         try {
             userDao.createUser(user);
-            return authDao.createAuth(user.username()); //maybe outside?
+            return authDao.createAuth(user.username());
         } catch (Exception e){
-            throw new DataAccessException("Failed to create user");
+            throw new ResponseException(500, "Error: " + e.getMessage());
         }
     }
 
-    public AuthData login(UserData user) throws DataAccessException {
+    public AuthData login(UserData user) throws ResponseException {
         if (user == null || user.username() == null || user.username().isEmpty() || user.password() == null || user.password().isEmpty()){
-            throw new DataAccessException("Not enough data provided");
+            throw new ResponseException(401, "Error: unauthorized");
         }
+
         UserData currentUser = userDao.getUser(user.username());
 
         if (currentUser == null || !currentUser.password().equals(user.password())){
-            throw new DataAccessException("Wrong user or password");
+            throw new ResponseException(401, "Error: unauthorized");
         }
 
         try {
             return authDao.createAuth(user.username());
         } catch (Exception e){
-            throw new DataAccessException("Failed to login");
+            throw new ResponseException(500, "Error: " + e.getMessage());
         }
     }
 
-    public void logout(String authToken) throws DataAccessException {
-        if (authToken == null || authToken.isEmpty()) {
-            throw new DataAccessException("Invalid auth token");
+    public void logout(String authToken) throws ResponseException {
+        if (authToken == null || authToken.isEmpty() || authDao.getAuth(authToken) == null) {
+            throw new ResponseException(401, "Error: unauthorized");
         }
         try {
             authDao.deleteAuth(authToken);
         }catch (Exception e){
-            throw new DataAccessException("Failed to logout");
+            throw new ResponseException(500, "Error: " + e.getMessage());
         }
 
     }
